+++ keyboards/sofle/info.json
@@ -1,0 +1,120 @@
+{
+    "keyboard_name": "Sofle",
+    "features": {
+        "swap_hands": true,
+        "bootmagic": true,
+        "extrakey": true,
+        "mousekey": true,
+        "nkro": true,
+        "encoder": true,
+        "oled": true,
+        "rgblight": false,
+        "rgb_matrix": false
+    },
+    "build": {
+        "lto": true
+    },
+    "usb": {
+        "vid": "0xFC32"
+    },
+    "diode_direction": "COL2ROW",
+    "qmk": {
+        "tap_keycode_delay": 10
+    },
+    "split": {
+        "enabled": true,
+        "soft_serial_pin": "D2",
+        "transport": {
+            "sync": {
+                "matrix_state": true
+            }
+        }
+    },
+    "ws2812": {
+        "pin": "D3"
+    },
+    "rgblight": {
+        "led_count": 72,
+        "split_count": [36, 36],
+        "sleep": true
+    },
+    "rgb_matrix": {
+        "driver": "ws2812",
+        "split_count": [36, 36],
+        "sleep": true,
+        "layout": [
+            {"x": 96, "y": 40, "flags": 8},
+            {"x": 16, "y": 20, "flags": 2},
+            {"x": 48, "y": 10, "flags": 2},
+            {"x": 80, "y": 18, "flags": 2},
+            {"x": 88, "y": 60, "flags": 2},
+            {"x": 56, "y": 57, "flags": 2},
+            {"x": 24, "y": 60, "flags": 2},
+            {"matrix": [4, 0], "x": 32, "y": 57, "flags": 4},
+            {"matrix": [3, 0], "x": 0, "y": 48, "flags": 4},
+            {"matrix": [2, 0], "x": 0, "y": 36, "flags": 4},
+            {"matrix": [1, 0], "x": 0, "y": 24, "flags": 4},
+            {"matrix": [0, 0], "x": 0, "y": 12, "flags": 4},
+            {"matrix": [0, 1], "x": 16, "y": 12, "flags": 4},
+            {"matrix": [1, 1], "x": 16, "y": 24, "flags": 4},
+            {"matrix": [2, 1], "x": 16, "y": 36, "flags": 4},
+            {"matrix": [3, 1], "x": 16, "y": 48, "flags": 4},
+            {"matrix": [4, 1], "x": 48, "y": 55, "flags": 4},
+            {"matrix": [4, 2], "x": 64, "y": 57, "flags": 4},
+            {"matrix": [3, 2], "x": 32, "y": 45, "flags": 4},
+            {"matrix": [2, 2], "x": 32, "y": 33, "flags": 4},
+            {"matrix": [1, 2], "x": 32, "y": 21, "flags": 4},
+            {"matrix": [0, 2], "x": 32, "y": 9, "flags": 4},
+            {"matrix": [0, 3], "x": 48, "y": 7, "flags": 4},
+            {"matrix": [1, 3], "x": 48, "y": 19, "flags": 4},
+            {"matrix": [2, 3], "x": 48, "y": 31, "flags": 4},
+            {"matrix": [3, 3], "x": 48, "y": 43, "flags": 4},
+            {"matrix": [4, 3], "x": 80, "y": 59, "flags": 4},
+            {"matrix": [4, 4], "x": 96, "y": 64, "flags": 4},
+            {"matrix": [3, 4], "x": 64, "y": 45, "flags": 4},
+            {"matrix": [2, 4], "x": 64, "y": 33, "flags": 4},
+            {"matrix": [1, 4], "x": 64, "y": 21, "flags": 4},
+            {"matrix": [0, 4], "x": 64, "y": 9, "flags": 4},
+            {"matrix": [0, 5], "x": 80, "y": 10, "flags": 4},
+            {"matrix": [1, 5], "x": 80, "y": 22, "flags": 4},
+            {"matrix": [2, 5], "x": 80, "y": 34, "flags": 4},
+            {"matrix": [3, 5], "x": 80, "y": 47, "flags": 4},
+            {"x": 128, "y": 40, "flags": 8},
+            {"x": 208, "y": 20, "flags": 2},
+            {"x": 176, "y": 10, "flags": 2},
+            {"x": 144, "y": 18, "flags": 2},
+            {"x": 136, "y": 60, "flags": 2},
+            {"x": 168, "y": 57, "flags": 2},
+            {"x": 200, "y": 60, "flags": 2},
+            {"matrix": [9, 0], "x": 192, "y": 57, "flags": 4},
+            {"matrix": [8, 0], "x": 224, "y": 48, "flags": 4},
+            {"matrix": [7, 0], "x": 224, "y": 36, "flags": 4},
+            {"matrix": [6, 0], "x": 224, "y": 24, "flags": 4},
+            {"matrix": [5, 0], "x": 224, "y": 12, "flags": 4},
+            {"matrix": [5, 1], "x": 208, "y": 12, "flags": 4},
+            {"matrix": [6, 1], "x": 208, "y": 24, "flags": 4},
+            {"matrix": [7, 1], "x": 208, "y": 36, "flags": 4},
+            {"matrix": [8, 1], "x": 208, "y": 48, "flags": 4},
+            {"matrix": [9, 1], "x": 176, "y": 55, "flags": 4},
+            {"matrix": [9, 2], "x": 160, "y": 57, "flags": 4},
+            {"matrix": [8, 2], "x": 192, "y": 45, "flags": 4},
+            {"matrix": [7, 2], "x": 192, "y": 33, "flags": 4},
+            {"matrix": [6, 2], "x": 192, "y": 21, "flags": 4},
+            {"matrix": [5, 2], "x": 192, "y": 9, "flags": 4},
+            {"matrix": [5, 3], "x": 176, "y": 7, "flags": 4},
+            {"matrix": [6, 3], "x": 176, "y": 19, "flags": 4},
+            {"matrix": [7, 3], "x": 176, "y": 31, "flags": 4},
+            {"matrix": [8, 3], "x": 176, "y": 43, "flags": 4},
+            {"matrix": [9, 3], "x": 144, "y": 59, "flags": 4},
+            {"matrix": [9, 4], "x": 128, "y": 64, "flags": 4},
+            {"matrix": [8, 4], "x": 160, "y": 45, "flags": 4},
+            {"matrix": [7, 4], "x": 160, "y": 33, "flags": 4},
+            {"matrix": [6, 4], "x": 160, "y": 21, "flags": 4},
+            {"matrix": [5, 4], "x": 160, "y": 9, "flags": 4},
+            {"matrix": [5, 5], "x": 144, "y": 10, "flags": 4},
+            {"matrix": [6, 5], "x": 144, "y": 22, "flags": 4},
+            {"matrix": [7, 5], "x": 144, "y": 34, "flags": 4},
+            {"matrix": [8, 5], "x": 144, "y": 47, "flags": 4}
+        ]
+    }
+}
