+++ keyboards/sofle/keymaps/default/rules.mk
@@ -1,5 +1,1 @@
-
-OLED_ENABLE = yes
-ENCODER_ENABLE = yes
-CONSOLE_ENABLE = no
-EXTRAKEY_ENABLE = yes
+TRI_LAYER_ENABLE = yes
