+++ keyboards/sofle/keymaps/via/rules.mk
@@ -1,8 +1,3 @@
-OLED_ENABLE = yes
-ENCODER_ENABLE = yes
-CONSOLE_ENABLE = no
-EXTRAKEY_ENABLE = yes
 VIA_ENABLE = yes
-LTO_ENABLE = yes
-RGBLIGHT_ENABLE = yes
+RGB_MATRIX_ENABLE = yes
 ENCODER_MAP_ENABLE = yes
