+++ keyboards/sofle/rev1/rules.mk
@@ -1,2 +1,1 @@
-ENCODER_ENABLE = yes
-OLED_ENABLE = yes
+# This file intentionally left blank
