+++ keyboards/sofle/rules.mk
@@ -1,3 +1,1 @@
-SPLIT_KEYBOARD = yes
 DEFAULT_FOLDER = sofle/rev1
-SWAP_HANDS_ENABLE = yes
